Modify Activity Name
