+++ app/src/main/java/woo/sopt22/meowbox/ApplicationController.kt
@@ -30,3 +30,4 @@ class ApplicationController : Application() {
         networkService = retrofit.create(NetworkService::class.java)
     }
 }
+
