+++ app/src/main/java/woo/sopt22/meowbox/View/Order/OrderFragment/OrderWithOutCatInfoActivity.kt
@@ -1,96 +1,88 @@
-package woo.sopt22.meowbox.View.Order
+package woo.sopt22.meowbox.View.Order.OrderFragment
 
-import android.content.Context
 import android.content.Intent
+import android.os.Bundle
+import android.support.design.widget.NavigationView
+import android.support.v4.view.GravityCompat
+import android.support.v7.app.ActionBarDrawerToggle
+import android.support.v7.app.AppCompatActivity
+import android.view.MenuItem
+import kotlinx.android.synthetic.main.activity_order_first.*
+import kotlinx.android.synthetic.main.app_bar_order_first.*
+import woo.sopt22.meowbox.R
+import woo.sopt22.meowbox.View.Home.MainActivity
+import woo.sopt22.meowbox.View.MeowBoxStory.MeowBoxStoryActivity
+import woo.sopt22.meowbox.View.MyPage.MyPageActivity
+import android.support.v4.content.res.ResourcesCompat
+import android.view.View
+import android.content.Context
 import android.graphics.Bitmap
 import android.graphics.Color
 import android.graphics.drawable.BitmapDrawable
 import android.graphics.drawable.ColorDrawable
 import android.net.Uri
 import android.os.Build
-import android.os.Bundle
-import android.support.design.widget.NavigationView
 import android.support.v4.app.Fragment
-import android.support.v4.content.res.ResourcesCompat
-import android.support.v4.view.GravityCompat
-import android.support.v7.app.ActionBarDrawerToggle
-import android.support.v7.app.AppCompatActivity
 import android.util.Log
 import android.view.Menu
-import android.view.MenuItem
-import android.view.View
+import android.widget.FrameLayout
 import android.widget.ImageView
+import android.widget.RelativeLayout
 import android.widget.TextView
 import com.bumptech.glide.Glide
-import kotlinx.android.synthetic.main.activity_order_third.*
-import kotlinx.android.synthetic.main.app_bar_order_third.*
-import woo.sopt22.meowbox.R
+import kotlinx.android.synthetic.main.content_order_first.*
 import woo.sopt22.meowbox.Util.CustomDialog.CatCustomDialog
 import woo.sopt22.meowbox.Util.SharedPreference
 import woo.sopt22.meowbox.Util.ToastMaker
-import woo.sopt22.meowbox.View.Home.MainActivity
 import woo.sopt22.meowbox.View.Login.LoginActivity
 import woo.sopt22.meowbox.View.MeowBoxReview.MeowBoxReviewActivity
-import woo.sopt22.meowbox.View.MeowBoxStory.MeowBoxStoryActivity
-import woo.sopt22.meowbox.View.MyPage.MyPageActivity
-import woo.sopt22.meowbox.View.Order.OrderFragmentWithCatInfo.WithCatInfoThird
+import woo.sopt22.meowbox.View.Order.LoginCustomDialog
 
-class OrderThirdActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelectedListener {
+
+class OrderWithOutCatInfoActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelectedListener, View.OnClickListener {
+    override fun onClick(v: View?) {
+        when(v!!){
+            order_next_btn->{
+                //startActivity(Intent(this, OrderSecondActivity::class.java))
+                val dialog = LoginCustomDialog(this)
+                dialog.show()
+                dialog.dismiss()
+            }
+        }
+    }
+
 
     companion object {
-        var thirdContext : Context?=null
+        var mContext: Context?=null
     }
 
     fun init(){
-
-        // Status Bar 색상 변경 및 하단 Navigation Bar 색상 변경
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP){
             window.statusBarColor = Color.BLACK
             window.navigationBarColor = Color.BLACK
         }
 
-        // 툴바 가리고 내가 원하는 이미지 넣을 수 있도록
-        getSupportActionBar()!!.setDisplayShowTitleEnabled(false)
-        getSupportActionBar()!!.setDisplayHomeAsUpEnabled(true)
-        toolbar.bringToFront()
-
-    }
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_order_third)
-        setSupportActionBar(toolbar)
-        thirdContext = this
-
-        init()
-        replaceFragment(WithCatInfoThird())
-
-
-        var headerView : View = order_third_nav_view.getHeaderView(0)
-        var userName : TextView = headerView.findViewById<TextView>(R.id.header_name)
-        var userImage : ImageView = headerView.findViewById(R.id.imageView)
         SharedPreference.instance!!.load(this)
 
-        var menu : Menu = order_third_nav_view.menu
-        var menu_item : MenuItem = menu.findItem(R.id.loginBtn)
-        var blank_menu_item : MenuItem = menu.findItem(R.id.blankBtn)
-        var blank_menu_item2 : MenuItem = menu.findItem(R.id.blankBtn2)
-        var order_btn : MenuItem = menu.findItem(R.id.orderBtn)
-        order_btn.setEnabled(false)
-        blank_menu_item.setEnabled(false)
-        blank_menu_item2.setEnabled(false)
+        getSupportActionBar()!!.setDisplayShowTitleEnabled(false)
+        getSupportActionBar()!!.setDisplayHomeAsUpEnabled(true)
+
+    }
+    lateinit var order_next_btn : RelativeLayout
+    lateinit var container : FrameLayout
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_order_first)
+        setSupportActionBar(toolbar)
+        mContext = this
+
+        init()
+        replaceFragment(OrderFirstFragment())
 
 
-        if(SharedPreference.instance!!.getPrefStringData("name")!!.isEmpty()){
-            userName.text = "OO님!"
-            menu_item.setTitle("로그인")
-
-        } else {
-            userName.text = SharedPreference.instance!!.getPrefStringData("name")
-            menu_item.setTitle("")
-            menu_item.setEnabled(false)
-        }
-
-
+        var headerView : View = order_first_nav_view.getHeaderView(0)
+        var userName : TextView = headerView.findViewById<TextView>(R.id.header_name)
+        var userImage : ImageView = headerView.findViewById(R.id.imageView)
 
         if(SharedPreference.instance!!.getPrefStringData("image_profile") == null){
             Log.v("용범 onResume","123")
@@ -101,8 +93,31 @@ class OrderThirdActivity : AppCompatActivity(), NavigationView.OnNavigationItemS
             Glide.with(this).load(SharedPreference.instance!!.getPrefStringData("image_profile")!!).into(userImage)
         }
 
+
+        container = order_framge as FrameLayout
+
+        // 네비게이션 버튼 비활성화
+        var menu : Menu = order_first_nav_view.menu
+        var menu_item : MenuItem = menu.findItem(R.id.loginBtn)
+        var blank_menu_item : MenuItem = menu.findItem(R.id.blankBtn)
+        var blank_menu_item2 : MenuItem = menu.findItem(R.id.blankBtn2)
+        var order_btn : MenuItem = menu.findItem(R.id.orderBtn)
+        order_btn.setEnabled(false)
+        blank_menu_item.setEnabled(false)
+        blank_menu_item2.setEnabled(false)
+
+        if(SharedPreference.instance!!.getPrefStringData("name")!!.isEmpty()){
+            userName.text = "OO님!"
+            menu_item.setTitle("로그인")
+        } else {
+            userName.text = SharedPreference.instance!!.getPrefStringData("name") + " 님"
+            menu_item.setTitle("")
+            menu_item.setEnabled(false)
+        }
+
+
         val toggle = ActionBarDrawerToggle(
-                this, order_third_drawer_layout, toolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close)
+                this, order_first_drawer_layout, toolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close)
 
         toggle.setDrawerIndicatorEnabled(false)
         val drawable = ResourcesCompat.getDrawable(resources, R.drawable.side_bar_btn_black, applicationContext!!.getTheme())
@@ -113,28 +128,37 @@ class OrderThirdActivity : AppCompatActivity(), NavigationView.OnNavigationItemS
         toggle.setHomeAsUpIndicator(drawable)
         toggle.setToolbarNavigationClickListener(object : View.OnClickListener {
             override fun onClick(v: View?) {
-                if (order_third_drawer_layout.isDrawerVisible(GravityCompat.START)) {
-                    order_third_drawer_layout.closeDrawer(GravityCompat.START)
+                if (order_first_drawer_layout.isDrawerVisible(GravityCompat.START)) {
+                    order_first_drawer_layout.closeDrawer(GravityCompat.START)
                 } else {
-                    order_third_drawer_layout.openDrawer(GravityCompat.START)
+                    order_first_drawer_layout.openDrawer(GravityCompat.START)
                 }
             }
         })
-        order_third_drawer_layout.addDrawerListener(toggle)
+
+        order_first_drawer_layout.addDrawerListener(toggle)
         toggle.syncState()
 
-        order_third_nav_view.setNavigationItemSelectedListener(this)
+        order_first_nav_view.setNavigationItemSelectedListener(this)
+    }
 
 
+    fun replaceFragment(fragment : Fragment){
+        val fm = supportFragmentManager
+        val transaction = fm.beginTransaction()
+        transaction.replace(R.id.order_framge, fragment)
+        transaction.commit()
     }
 
     override fun onBackPressed() {
-        if (order_third_drawer_layout.isDrawerOpen(GravityCompat.START)) {
-            order_third_drawer_layout.closeDrawer(GravityCompat.START)
+        if (order_first_drawer_layout.isDrawerOpen(GravityCompat.START)) {
+            order_first_drawer_layout.closeDrawer(GravityCompat.START)
         } else {
             super.onBackPressed()
         }
     }
+
+
 
 
     override fun onNavigationItemSelected(item: MenuItem): Boolean {
@@ -151,7 +175,7 @@ class OrderThirdActivity : AppCompatActivity(), NavigationView.OnNavigationItemS
                 item.isChecked = false
             }
             R.id.homeBtn -> {
-                startActivity(Intent(thirdContext, MainActivity::class.java))
+                startActivity(Intent(mContext, MainActivity::class.java))
             }
             R.id.stroyBtn -> {
                 startActivity(Intent(this, MeowBoxStoryActivity::class.java))
@@ -167,7 +191,7 @@ class OrderThirdActivity : AppCompatActivity(), NavigationView.OnNavigationItemS
                         dialog.window!!.setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT))
                         dialog.show()
                     } else {
-                        val intent = Intent(this, OrderFirstActivity::class.java)
+                        val intent = Intent(this, OrderWithOutCatInfoActivity::class.java)
                         intent.putExtra("cat_idx", SharedPreference.instance!!.getPrefStringData("cat_idx")!!)
                         startActivity(intent)
                     }
@@ -186,13 +210,7 @@ class OrderThirdActivity : AppCompatActivity(), NavigationView.OnNavigationItemS
             }
         }
 
-        order_third_drawer_layout.closeDrawer(GravityCompat.START)
+        order_first_drawer_layout.closeDrawer(GravityCompat.START)
         return true
-    }
-    fun replaceFragment(fragment : Fragment){
-        val fm = supportFragmentManager
-        val transaction = fm.beginTransaction()
-        transaction.replace(R.id.order_framge, fragment)
-        transaction.commit()
     }
 }
