+++ app/src/main/java/woo/sopt22/meowbox/View/Order/OrderFragmentWithCatInfo/OrderWithCatInfoActivity.kt
@@ -1,90 +1,97 @@
-package woo.sopt22.meowbox.View.Order
+package woo.sopt22.meowbox.View.Order.OrderFragmentWithCatInfo
 
-import android.content.Intent
-import android.os.Bundle
-import android.support.design.widget.NavigationView
-import android.support.v4.view.GravityCompat
-import android.support.v7.app.ActionBarDrawerToggle
-import android.support.v7.app.AppCompatActivity
-import android.view.MenuItem
-import kotlinx.android.synthetic.main.activity_order_first.*
-import kotlinx.android.synthetic.main.app_bar_order_first.*
-import woo.sopt22.meowbox.R
-import woo.sopt22.meowbox.View.Home.MainActivity
-import woo.sopt22.meowbox.View.MeowBoxStory.MeowBoxStoryActivity
-import woo.sopt22.meowbox.View.MyPage.MyPageActivity
-import android.support.v4.content.res.ResourcesCompat
-import android.view.View
 import android.content.Context
+import android.content.Intent
 import android.graphics.Bitmap
 import android.graphics.Color
 import android.graphics.drawable.BitmapDrawable
 import android.graphics.drawable.ColorDrawable
 import android.net.Uri
 import android.os.Build
+import android.os.Bundle
+import android.support.design.widget.NavigationView
 import android.support.v4.app.Fragment
+import android.support.v4.content.res.ResourcesCompat
+import android.support.v4.view.GravityCompat
+import android.support.v7.app.ActionBarDrawerToggle
+import android.support.v7.app.AppCompatActivity
 import android.util.Log
 import android.view.Menu
-import android.widget.FrameLayout
+import android.view.MenuItem
+import android.view.View
 import android.widget.ImageView
-import android.widget.RelativeLayout
 import android.widget.TextView
 import com.bumptech.glide.Glide
 import kotlinx.android.synthetic.main.activity_order_third.*
-import kotlinx.android.synthetic.main.content_order_first.*
+import kotlinx.android.synthetic.main.app_bar_order_third.*
+import woo.sopt22.meowbox.R
 import woo.sopt22.meowbox.Util.CustomDialog.CatCustomDialog
 import woo.sopt22.meowbox.Util.SharedPreference
 import woo.sopt22.meowbox.Util.ToastMaker
+import woo.sopt22.meowbox.View.Home.MainActivity
 import woo.sopt22.meowbox.View.Login.LoginActivity
 import woo.sopt22.meowbox.View.MeowBoxReview.MeowBoxReviewActivity
-import woo.sopt22.meowbox.View.Order.OrderFragment.OrderFirstFragment
-import woo.sopt22.meowbox.View.Order.OrderFragment.OrderThirdFragment
+import woo.sopt22.meowbox.View.MeowBoxStory.MeowBoxStoryActivity
+import woo.sopt22.meowbox.View.MyPage.MyPageActivity
+import woo.sopt22.meowbox.View.Order.LoginCustomDialog
+import woo.sopt22.meowbox.View.Order.OrderFragment.OrderWithOutCatInfoActivity
 
-
-class OrderFirstActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelectedListener, View.OnClickListener {
-    override fun onClick(v: View?) {
-        when(v!!){
-            order_next_btn->{
-                //startActivity(Intent(this, OrderSecondActivity::class.java))
-                val dialog = LoginCustomDialog(this)
-                dialog.show()
-                dialog.dismiss()
-            }
-        }
-    }
-
+class OrderWithCatInfoActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelectedListener {
 
     companion object {
-        var mContext: Context?=null
+        var thirdContext : Context?=null
     }
 
     fun init(){
+
+        // Status Bar 색상 변경 및 하단 Navigation Bar 색상 변경
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP){
             window.statusBarColor = Color.BLACK
             window.navigationBarColor = Color.BLACK
         }
 
-        SharedPreference.instance!!.load(this)
-
+        // 툴바 가리고 내가 원하는 이미지 넣을 수 있도록
         getSupportActionBar()!!.setDisplayShowTitleEnabled(false)
         getSupportActionBar()!!.setDisplayHomeAsUpEnabled(true)
+        toolbar.bringToFront()
 
     }
-    lateinit var order_next_btn : RelativeLayout
-    lateinit var container : FrameLayout
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_order_first)
+        setContentView(R.layout.activity_order_third)
         setSupportActionBar(toolbar)
-        mContext = this
+        thirdContext = this
 
         init()
-        replaceFragment(OrderFirstFragment())
+        replaceFragment(WithCatInfoThird())
 
 
-        var headerView : View = order_first_nav_view.getHeaderView(0)
+        var headerView : View = order_third_nav_view.getHeaderView(0)
         var userName : TextView = headerView.findViewById<TextView>(R.id.header_name)
         var userImage : ImageView = headerView.findViewById(R.id.imageView)
+        SharedPreference.instance!!.load(this)
+
+        var menu : Menu = order_third_nav_view.menu
+        var menu_item : MenuItem = menu.findItem(R.id.loginBtn)
+        var blank_menu_item : MenuItem = menu.findItem(R.id.blankBtn)
+        var blank_menu_item2 : MenuItem = menu.findItem(R.id.blankBtn2)
+        var order_btn : MenuItem = menu.findItem(R.id.orderBtn)
+        order_btn.setEnabled(false)
+        blank_menu_item.setEnabled(false)
+        blank_menu_item2.setEnabled(false)
+
+
+        if(SharedPreference.instance!!.getPrefStringData("name")!!.isEmpty()){
+            userName.text = "OO님!"
+            menu_item.setTitle("로그인")
+
+        } else {
+            userName.text = SharedPreference.instance!!.getPrefStringData("name")
+            menu_item.setTitle("")
+            menu_item.setEnabled(false)
+        }
+
+
 
         if(SharedPreference.instance!!.getPrefStringData("image_profile") == null){
             Log.v("용범 onResume","123")
@@ -95,31 +102,8 @@ class OrderFirstActivity : AppCompatActivity(), NavigationView.OnNavigationItemS
             Glide.with(this).load(SharedPreference.instance!!.getPrefStringData("image_profile")!!).into(userImage)
         }
 
-
-        container = order_framge as FrameLayout
-
-        // 네비게이션 버튼 비활성화
-        var menu : Menu = order_first_nav_view.menu
-        var menu_item : MenuItem = menu.findItem(R.id.loginBtn)
-        var blank_menu_item : MenuItem = menu.findItem(R.id.blankBtn)
-        var blank_menu_item2 : MenuItem = menu.findItem(R.id.blankBtn2)
-        var order_btn : MenuItem = menu.findItem(R.id.orderBtn)
-        order_btn.setEnabled(false)
-        blank_menu_item.setEnabled(false)
-        blank_menu_item2.setEnabled(false)
-
-        if(SharedPreference.instance!!.getPrefStringData("name")!!.isEmpty()){
-            userName.text = "OO님!"
-            menu_item.setTitle("로그인")
-        } else {
-            userName.text = SharedPreference.instance!!.getPrefStringData("name") + " 님"
-            menu_item.setTitle("")
-            menu_item.setEnabled(false)
-        }
-
-
         val toggle = ActionBarDrawerToggle(
-                this, order_first_drawer_layout, toolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close)
+                this, order_third_drawer_layout, toolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close)
 
         toggle.setDrawerIndicatorEnabled(false)
         val drawable = ResourcesCompat.getDrawable(resources, R.drawable.side_bar_btn_black, applicationContext!!.getTheme())
@@ -130,37 +114,28 @@ class OrderFirstActivity : AppCompatActivity(), NavigationView.OnNavigationItemS
         toggle.setHomeAsUpIndicator(drawable)
         toggle.setToolbarNavigationClickListener(object : View.OnClickListener {
             override fun onClick(v: View?) {
-                if (order_first_drawer_layout.isDrawerVisible(GravityCompat.START)) {
-                    order_first_drawer_layout.closeDrawer(GravityCompat.START)
+                if (order_third_drawer_layout.isDrawerVisible(GravityCompat.START)) {
+                    order_third_drawer_layout.closeDrawer(GravityCompat.START)
                 } else {
-                    order_first_drawer_layout.openDrawer(GravityCompat.START)
+                    order_third_drawer_layout.openDrawer(GravityCompat.START)
                 }
             }
         })
-
-        order_first_drawer_layout.addDrawerListener(toggle)
+        order_third_drawer_layout.addDrawerListener(toggle)
         toggle.syncState()
 
-        order_first_nav_view.setNavigationItemSelectedListener(this)
-    }
+        order_third_nav_view.setNavigationItemSelectedListener(this)
 
 
-    fun replaceFragment(fragment : Fragment){
-        val fm = supportFragmentManager
-        val transaction = fm.beginTransaction()
-        transaction.replace(R.id.order_framge, fragment)
-        transaction.commit()
     }
 
     override fun onBackPressed() {
-        if (order_first_drawer_layout.isDrawerOpen(GravityCompat.START)) {
-            order_first_drawer_layout.closeDrawer(GravityCompat.START)
+        if (order_third_drawer_layout.isDrawerOpen(GravityCompat.START)) {
+            order_third_drawer_layout.closeDrawer(GravityCompat.START)
         } else {
             super.onBackPressed()
         }
     }
-
-
 
 
     override fun onNavigationItemSelected(item: MenuItem): Boolean {
@@ -177,7 +152,7 @@ class OrderFirstActivity : AppCompatActivity(), NavigationView.OnNavigationItemS
                 item.isChecked = false
             }
             R.id.homeBtn -> {
-                startActivity(Intent(OrderFirstActivity.mContext, MainActivity::class.java))
+                startActivity(Intent(thirdContext, MainActivity::class.java))
             }
             R.id.stroyBtn -> {
                 startActivity(Intent(this, MeowBoxStoryActivity::class.java))
@@ -193,7 +168,7 @@ class OrderFirstActivity : AppCompatActivity(), NavigationView.OnNavigationItemS
                         dialog.window!!.setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT))
                         dialog.show()
                     } else {
-                        val intent = Intent(this, OrderFirstActivity::class.java)
+                        val intent = Intent(this, OrderWithOutCatInfoActivity::class.java)
                         intent.putExtra("cat_idx", SharedPreference.instance!!.getPrefStringData("cat_idx")!!)
                         startActivity(intent)
                     }
@@ -212,7 +187,13 @@ class OrderFirstActivity : AppCompatActivity(), NavigationView.OnNavigationItemS
             }
         }
 
-        order_first_drawer_layout.closeDrawer(GravityCompat.START)
+        order_third_drawer_layout.closeDrawer(GravityCompat.START)
         return true
+    }
+    fun replaceFragment(fragment : Fragment){
+        val fm = supportFragmentManager
+        val transaction = fm.beginTransaction()
+        transaction.replace(R.id.order_framge, fragment)
+        transaction.commit()
     }
 }
