+++ app/src/main/res/layout/activity_my_page_setting.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".View.MyPage.Setting.MyPageSettingActivity"
+    tools:context=".View.MyPage.Setting.SettingActivity"
     android:orientation="vertical">
 
     <LinearLayout
