+++ app/src/main/res/layout/activity_my_setting.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".View.MyPage.MySettingActivity">
+    tools:context=".View.MyPage.MyPageSettingActivity">
 
     <ScrollView
         android:overScrollMode="never"
