+++ app/src/main/res/layout/app_bar_order_first.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".View.Order.OrderFirstActivity"
+    tools:context=".View.Order.OrderFragment.OrderWithOutCatInfoActivity"
     android:background="#FFFFFF">
 
     <android.support.design.widget.AppBarLayout
