+++ app/src/main/res/layout/app_bar_order_third.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".View.Order.OrderThirdActivity">
+    tools:context=".View.Order.OrderFragmentWithCatInfo.OrderWithCatInfoActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
