+++ app/src/main/res/layout/content_order_first.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".View.Order.OrderFirstActivity"
+    tools:context=".View.Order.OrderFragment.OrderWithOutCatInfoActivity"
     tools:showIn="@layout/app_bar_order_first">
 
     <FrameLayout
