+++ app/src/main/res/layout/content_order_third.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".View.Order.OrderThirdActivity"
+    tools:context=".View.Order.OrderFragmentWithCatInfo.OrderWithCatInfoActivity"
     tools:showIn="@layout/app_bar_order_third">
 
     <FrameLayout
